# business/openapi.yaml
openapi: 3.1.0
info:
  title: rust-invoice-system — business API
  version: "1.0.0"
  description: |
    Invoicing, charging and event delivery.

    ## Two audiences, one port

    Which endpoints need a token is decided by **who is asking**, not by whether the request
    writes. `GET /invoices` is gated and `POST /invoices/{id}/pay` is not, because raising or
    voiding an invoice is the business's own affair, while whoever is being billed has no token
    and cannot be given one.

    - **Operator** — the business's back office. Requires `X-API-Token`.
    - **Payer** — the person being billed. `POST /invoices/{id}/pay` and
      `GET /payment_intents/{id}`, both open by necessity. Neither exposes anything about an
      invoice beyond the charge the caller is already making.
    - **Docker** — `GET /health`, open so the container healthcheck can reach it.

    An unauthenticated request to an operator endpoint gets `401 unauthorized`; a *wrong* token
    gets exactly the same, since telling that caller their header was well formed and merely
    incorrect hands them the one bit they were missing. A path matching no route stays `404`
    rather than `401`, which would reveal which paths exist.

    ## Error format

    Every failure answers with the same shape, so a caller parses failures one way:

    ```json
    { "error": "invoice_not_payable" }
    ```

    One addition, on `POST /invoices/{id}/pay`: once a charge has reached the payment service, every
...

...

...

...
servers:
...
paths:
...
  /customers/{id}:
    get:
      tags: [Customers]
      summary: Read one customer
      operationId: getCustomer
      security: [{ operatorToken: [] }]
      parameters: [{ $ref: "#/components/parameters/CustomerId" }]
      responses:
        "200":
          description: The customer.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Customer" }
        "400":
          description: "`invalid_customer_id` — not a uuid."
          $ref: "#/components/responses/Error"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404":
          description: "`customer_not_found`"
          $ref: "#/components/responses/Error"
        "500": { $ref: "#/components/responses/Internal" }
  /invoices:
    post:
      tags: [Invoices]
      summary: Raise an invoice
      operationId: createInvoice
      description: |
        The caller sends line items and the server totals them; there is no way to supply a total
        directly, so the number stored can never disagree with the lines it came from. The line
        items themselves are **not persisted** — they are input to the sum and then discarded.

        A new invoice is always `draft`. Emits `invoice.created`.
      security: [{ operatorToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateInvoice" }
            example:
              customer_id: "01a04ca6-d4e4-7e81-b6dc-41d699f8fa25"
              due_date: "2026-10-01"
              line_items:
                - { description: "Consulting", quantity: 2, unit_amount_cents: 50000 }
                - { description: "Support", quantity: 1, unit_amount_cents: 2500 }
      responses:
        "201":
          description: Created, in state `draft`.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Invoice" }
        "400":
          description: |
            `no_line_items` — the array is empty.
            `invalid_line_item` — blank description, `quantity <= 0`, or a negative unit amount.
            `total_too_large` — the sum overflows a 64-bit integer.
            `invalid_due_date` — not `YYYY-MM-DD`, or not a real calendar day.
            `invalid_customer_id` — not a uuid.
          $ref: "#/components/responses/Error"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404":
          description: "`customer_not_found` — the body is well formed, the customer is not there."
          $ref: "#/components/responses/Error"
        "500": { $ref: "#/components/responses/Internal" }
    get:
      tags: [Invoices]
      summary: List invoices
      operationId: listInvoices
      security: [{ operatorToken: [] }]
      parameters:
        - name: state
          in: query
          required: false
          description: Return only invoices in this state. Omit for all of them.
          schema: { $ref: "#/components/schemas/InvoiceState" }
      responses:
        "200":
          description: Matching invoices, ordered by id.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Invoice" }
        "400":
          description: "`invalid_invoice_state` — not one of the five states."
          $ref: "#/components/responses/Error"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/Internal" }
...
  /invoices/{id}/pay:
    post:
      tags: [Payments]
      summary: Charge a ready invoice
      operationId: payInvoice
      security: []
      description: |
        **No `X-API-Token`** — this is the payer's endpoint.

        **`Idempotency-Key` is required, not optional.** A charge with no key is a charge nobody
        can safely retry, and letting one through would make the protection depend on the caller
        remembering to ask for it, on the one endpoint where forgetting bills someone twice.

        The guarantee is exact and has no expiry: **one key, one response, forever.** The first
        request under a key is charged and its answer recorded; every later request under that
        key is handed that same answer back verbatim, whatever its status. A caller who wants a
        genuinely new attempt uses a new key.

        Concurrency is settled by a row lock whose state test is inside the locking statement, so
        of N simultaneous charges of one invoice **at most one is charged** and the rest get
        `409 invoice_not_payable`.

        **An unresolved charge is left unresolved.** If the payment service never answers, the
        invoice stays `processing` still pointing at its intent and nothing is written, because
        the card may already have been charged. Follow `GET /payment_intents/{id}`; a daily
        reconciler closes it out.

        Emits `invoice.paid` or `invoice.payment_failed` once the outcome is recorded.
      parameters:
        - $ref: "#/components/parameters/InvoiceId"
        - name: Idempotency-Key
          in: header
          required: true
          description: Any non-empty string, unique per intended charge.
          schema: { type: string, minLength: 1 }
          example: acme-oct-01
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PayInvoice" }
            example: { card_token: "tok_success" }
      responses:
        "200":
          description: Charged. The invoice is now `processed`.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Attempt" }
        "400":
          description: |
            `idempotency_key_missing` — absent or empty header.
            `idempotency_key_reused` — this key was already used for a *different* request. A key
            stands for one request; replaying would answer a question the caller did not ask.
            `invalid_invoice_id` — not a uuid.
          $ref: "#/components/responses/Error"
        "402":
          description: |
            `charge_declined` — the payment service declined the card. The invoice returns to
            `ready` and can be charged again with another card. Carries the intent id alongside
            the error.

            **Why is not reported.** The payment service separates `insufficient_funds` from
            `card_declined`; this endpoint answers `charge_declined` to both. The reason belongs
            between the cardholder and their bank, and putting the payment service's failure
            vocabulary into this API would make every caller depend on it.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AttemptError" }
        "404":
          description: "`invoice_not_found`"
          $ref: "#/components/responses/Error"
        "409":
          description: |
            `invoice_not_payable` — the invoice is not `ready`: still a draft, already paid,
            voided, or being charged by someone else right now.
            `idempotency_key_in_flight` — the first request under this key has not finished.
            Retry the same key once it has.
          $ref: "#/components/responses/Error"
        "413":
          description: "`request_body_too_large` — over 64 KiB."
          $ref: "#/components/responses/Error"
        "500":
          description: |
            This service's own database failed. **Carries both** `error` and
            `payment_intent_id`: the charge itself may well have succeeded, and the id is how the
            caller watches the record catch up.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AttemptError" }
        "502":
          description: "`payment_service_unavailable` — the PSP could not be reached. Nothing was charged and no invoice moved."
          $ref: "#/components/responses/Error"
        "504":
          description: |
            `charge_unresolved` — the charge reached the payment service and no answer came back.
            **The card may or may not have been charged**, and the code says only that: not that
            the charge failed, but that this service cannot say either way. The invoice is held
            in `processing`, and following the intent id is the only way to learn the outcome.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AttemptError" }
  /payment_intents/{id}:
    get:
      tags: [Payments]
      summary: What became of one charge attempt
      operationId: getPaymentIntent
      security: []
      description: |
        **No `X-API-Token`** — the other half of `POST /invoices/{id}/pay`, for a caller that got
        a `504` and has no idea whether the card was charged.

        **The recorded row is the answer.** No call to the payment service, no settlement, no
        side effect of any kind — that row is what the rest of the system acted on. The cost is
        that `pending` means *this service does not know yet*, not *the card was not charged*; a
        charge whose response was lost reads `pending` until the daily reconciler settles it.
      parameters:
        - name: id
          in: path
          required: true
          description: The payment service's own id.
          schema: { type: string }
          example: pi_55e3caceeccd43fa80cde32d642ec100
      responses:
        "200":
          description: The recorded status.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaymentIntent" }
        "404":
          description: "`payment_intent_not_found` — this service has no record of that intent."
          $ref: "#/components/responses/Error"
        "500": { $ref: "#/components/responses/Internal" }
...
components:
  securitySchemes:
    operatorToken:
      type: apiKey
      in: header
      name: X-API-Token
      description: |
        One shared static token, compared in constant time. A missing token and a wrong one get
        the same `401`.
  parameters:
    CustomerId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    InvoiceId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    Error:
      description: A failure, in the standard shape.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: "`unauthorized` — no token, or the wrong one."
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: unauthorized }
    Internal:
      description: "`internal_error`"
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: internal_error }
...
  schemas:
    Error:
      type: object
      description: The shape every failure answers with, bar the two documented exceptions.
      properties:
        error:
          type: string
          description: A stable machine-readable code. Never prose.
          example: invoice_not_payable
      required: [error]
...
    Customer:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        email: { type: string }
      required: [id, name, email]
    InvoiceState:
      type: string
      enum: [draft, ready, processing, processed, void]
      description: |
        `draft` — being prepared, editable.
        `ready` — finalized and payable.
        `processing` — a charge is in flight; no operator endpoint can move it.
        `processed` — paid. Terminal.
        `void` — written off. Terminal.
    LineItem:
      type: object
      description: Input to the total, then discarded. Not persisted and never returned.
      properties:
        description: { type: string, minLength: 1 }
        quantity:
          type: integer
          minimum: 1
          description: Must be positive — a zero-quantity line says nothing and is likelier a bug.
        unit_amount_cents:
          type: integer
          minimum: 0
          description: Zero is allowed; a comped line still belongs on the invoice.
      required: [description, quantity, unit_amount_cents]
    CreateInvoice:
      type: object
      properties:
        customer_id: { type: string, format: uuid }
        due_date:
          type: string
          format: date
          description: Exactly `YYYY-MM-DD`. Other formats are refused rather than guessed at.
        line_items:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/LineItem" }
      required: [customer_id, due_date, line_items]
    Invoice:
      type: object
      properties:
        id: { type: string, format: uuid }
        customer_id: { type: string, format: uuid }
        total_cents:
          type: integer
          minimum: 0
          description: Minor units, summed from the line items by the server.
        due_date: { type: string, format: date }
        state: { $ref: "#/components/schemas/InvoiceState" }
      required: [id, customer_id, total_cents, due_date, state]
    PayInvoice:
      type: object
      properties:
        card_token:
          type: string
          description: |
            A payment service card token. The development stack accepts `tok_success`,
            `tok_insufficient_funds`, `tok_card_declined`, `tok_timeout` and `tok_network_error`.
      required: [card_token]
    Attempt:
      type: object
      description: |
        The answer to a charge that reached the payment service: the intent id, and nothing else.
        Not the invoice — a caller that wants the invoice can read it; what it cannot get anywhere
        else is the handle to the attempt.
      properties:
        payment_intent_id: { type: string, example: pi_55e3caceeccd43fa80cde32d642ec100 }
      required: [payment_intent_id]
    AttemptError:
      type: object
      description: |
        Any failed charge that reached the payment service: the `402`, the `504` and the `500`.
        Carries the handle as well as the failure.
      properties:
        error: { type: string, enum: [charge_declined, charge_unresolved, internal_error] }
        payment_intent_id: { type: string }
      required: [error, payment_intent_id]
    PaymentStatus:
      type: string
      enum: [pending, succeeded, failed]
      description: |
        `pending` means **this service does not know yet**, not that the card was not charged.
    PaymentIntent:
      type: object
      properties:
        payment_intent_id: { type: string }
        status: { $ref: "#/components/schemas/PaymentStatus" }
      required: [payment_intent_id, status]
...